requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
billiard==3.6.4.0
celery==5.0.5
certifi==2023.11.17
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.103.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.4
mongoengine==0.27.0
mongomock==4.1.2
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.43
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
pytest==7.4.4
pytest-env==1.1.3
pytest-mock==3.12.0
python-dotenv==1.0.0
pytz==2023.3.post1
sentinels==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.23.2
vine==5.1.0
wcwidth==0.2.13
